Chore: 패키지 변경
- com.example.member => com.zoopick.server
