--- src/main/java/com/example/member/dto/CheckCertificationRequest.java
@@ -1,19 +1,0 @@
-package com.example.member.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class CheckCertificationRequest {
-
-    @NotBlank
-    @Email
-    private String email;
-    @NotBlank
-    private String certificationNumber;
-}
