--- src/main/java/com/example/member/dto/EmailcertificationRequest.java
@@ -1,18 +1,0 @@
-package com.example.member.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class EmailcertificationRequest {
-    
-    @NotBlank
-    @Email
-    private String email;
-
-}
