--- src/main/java/com/example/member/dto/FcmTokenRegistrationRequest.java
@@ -1,12 +1,0 @@
-package com.example.member.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class FcmTokenRegistrationRequest {
-    @NotBlank
-    private String token;
-}
