--- src/main/java/com/example/member/dto/LoginRequest.java
@@ -1,14 +1,0 @@
-package com.example.member.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class LoginRequest {
-    @NotBlank
-    private String schoolEmail;
-    @NotBlank
-    private String password;
-}
