--- src/main/java/com/example/member/dto/SignupRequest.java
@@ -1,14 +1,0 @@
-package com.example.member.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SignupRequest{
-    private String schoolEmail;
-    private String password;
-    private String nickname;
-    private String department;
-    private String grade;
-}
