--- src/main/java/com/example/member/entity/EmailAuth.java
@@ -1,20 +1,0 @@
-package com.example.member.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import lombok.*;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class EmailAuth {
-
-    @Id
-    private String email;
-
-    private String certificationNumber;
-
-    private boolean isVerified = false;
-}
