--- src/main/java/com/example/member/service/NotificationService.java
@@ -1,44 +1,0 @@
-package com.example.member.service;
-
-import com.example.member.dto.FcmNotificationRequest;
-import com.example.member.entity.User;
-import com.example.member.repository.UserRepository;
-import com.example.member.util.JwtUtil;
-import com.google.firebase.messaging.FirebaseMessaging;
-import com.google.firebase.messaging.FirebaseMessagingException;
-import com.google.firebase.messaging.Message;
-import com.google.firebase.messaging.Notification;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class NotificationService {
-    private final JwtUtil jwtUtil;
-    private final UserRepository userRepository;
-
-    public void register(String sessionToken, String fcmToken) {
-        String email = jwtUtil.extractEmail(sessionToken);
-        if (email == null)
-            throw new IllegalStateException("세션이 유효하지 않습니다.");
-        User user = userRepository.findBySchoolEmail(email)
-                .orElseThrow(() -> new IllegalStateException("사용자를 찾을 수 없습니다. | " + email));
-        user.setFcmToken(fcmToken);
-        userRepository.save(user);
-    }
-
-    public String send(String targetNickname, FcmNotificationRequest request) throws FirebaseMessagingException {
-        User user = userRepository.findByNickname(targetNickname)
-                .orElseThrow(() -> new IllegalStateException("사용자를 찾을 수 없습니다. | " + targetNickname));
-        Notification notification = Notification.builder()
-                .setTitle(request.getTitle())
-                .setBody(request.getBody())
-                .build();
-        Message message = Message.builder()
-                .setNotification(notification)
-                .putAllData(request.getData())
-                .setToken(user.getFcmToken())
-                .build();
-        return FirebaseMessaging.getInstance().send(message);
-    }
-}
