+++ src/main/java/com/zoopick/server/config/FirebaseConfig.java
@@ -1,0 +1,35 @@
+package com.zoopick.server.config;
+
+import com.google.auth.oauth2.GoogleCredentials;
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.FirebaseOptions;
+import jakarta.annotation.PostConstruct;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.springframework.context.annotation.Configuration;
+
+import java.io.FileInputStream;
+import java.io.IOException;
+
+@Configuration
+public class FirebaseConfig {
+    private static final Log LOGGER = LogFactory.getLog(FirebaseConfig.class);
+
+    @PostConstruct
+    public void setupFirebase() {
+        if (!System.getenv().containsKey("FIREBASE_ACCOUNT_KEY_PATH"))
+            throw new IllegalStateException("Environment variable FIREBASE_ACCOUNT_KEY_PATH is not set!");
+        try {
+            String accountKeyPath = System.getenv("FIREBASE_ACCOUNT_KEY_PATH");
+            FileInputStream serviceAccount = new FileInputStream(accountKeyPath);
+
+            FirebaseOptions options = FirebaseOptions.builder()
+                    .setCredentials(GoogleCredentials.fromStream(serviceAccount))
+                    .build();
+
+            FirebaseApp.initializeApp(options);
+        } catch (IOException e) {
+            LOGGER.error(e.getMessage());
+        }
+    }
+}
