+++ src/main/java/com/zoopick/server/dto/CheckCertificationRequest.java
@@ -1,0 +1,19 @@
+package com.zoopick.server.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class CheckCertificationRequest {
+
+    @NotBlank
+    @Email
+    private String email;
+    @NotBlank
+    private String certificationNumber;
+}
