+++ src/main/java/com/zoopick/server/dto/EmailcertificationRequest.java
@@ -1,0 +1,18 @@
+package com.zoopick.server.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class EmailcertificationRequest {
+    
+    @NotBlank
+    @Email
+    private String email;
+
+}
