+++ src/main/java/com/zoopick/server/dto/FcmTokenRegistrationRequest.java
@@ -1,0 +1,12 @@
+package com.zoopick.server.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class FcmTokenRegistrationRequest {
+    @NotBlank
+    private String token;
+}
