+++ src/main/java/com/zoopick/server/dto/LoginRequest.java
@@ -1,0 +1,14 @@
+package com.zoopick.server.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class LoginRequest {
+    @NotBlank
+    private String schoolEmail;
+    @NotBlank
+    private String password;
+}
