+++ src/main/java/com/zoopick/server/dto/SignupRequest.java
@@ -1,0 +1,14 @@
+package com.zoopick.server.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class SignupRequest{
+    private String schoolEmail;
+    private String password;
+    private String nickname;
+    private String department;
+    private String grade;
+}
