+++ src/main/java/com/zoopick/server/entity/EmailAuth.java
@@ -1,0 +1,23 @@
+package com.zoopick.server.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class EmailAuth {
+
+    @Id
+    private String email;
+
+    private String certificationNumber;
+
+    private boolean isVerified = false;
+}
