+++ src/main/java/com/zoopick/server/entity/User.java
@@ -1,0 +1,39 @@
+package com.zoopick.server.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Table(name = "users")
+@Getter
+@Setter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Builder
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "school_email", nullable = false, unique = true)
+    private String schoolEmail;
+
+    @Column
+    private String password;
+
+    @Column
+    private String nickname;
+
+    @Column(name = "fcm_token")
+    private String fcmToken;
+
+    @Column
+    private String role;
+
+    @Column
+    private String department;
+
+    @Column
+    private String grade;
+}
