+++ src/main/java/com/zoopick/server/repository/EmailAuthRepository.java
@@ -1,0 +1,9 @@
+package com.zoopick.server.repository;
+
+import com.zoopick.server.entity.EmailAuth;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface EmailAuthRepository extends JpaRepository<EmailAuth, String> {
+}
