+++ src/main/java/com/zoopick/server/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.zoopick.server.repository;
+
+import com.zoopick.server.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findBySchoolEmail(String schoolEmail);
+    Optional<User> findByNickname(String nickname);
+}
