+++ src/main/java/com/zoopick/server/util/JwtUtil.java
@@ -1,0 +1,55 @@
+package com.zoopick.server.util;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import org.springframework.stereotype.Component;
+
+import javax.crypto.SecretKey;
+import java.util.Date;
+
+@Component
+public class JwtUtil {
+    // 토큰 만료 시간
+    private static final long EXPIRATION_TIME = 1000L * 60 * 60 * 24;
+
+    // 1. 보안 키 생성
+    private final SecretKey secretKey = Jwts.SIG.HS256.key().build();
+
+    // 2. 토큰 생성
+    public String generateToken(String email) {
+        return Jwts.builder()
+                .subject(email)
+                .issuedAt(new Date())
+                .expiration(new Date(System.currentTimeMillis() + EXPIRATION_TIME))
+                .signWith(secretKey)
+                .compact();
+    }
+
+    // 3. 토큰에서 이메일 추출
+    public String extractEmail(String token) {
+        return getClaims(token).getSubject();
+    }
+
+    // 4. 토큰 유효성 검증
+    public boolean validateToken(String token) {
+        try {
+            return !getClaims(token).getExpiration().before(new Date());
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
+    private Claims getClaims(String token) {
+        return Jwts.parser()
+                .verifyWith(secretKey)
+                .build()
+                .parseSignedClaims(token)
+                .getPayload();
+    }
+
+    public long getExpiration(String token) {
+        Date expiration = getClaims(token).getExpiration();
+        long now = new Date().getTime();
+        return (expiration.getTime() - now);
+    }
+}
